refactor(agent-core-v2): own the engine version in _base/version.ts
Per review: a telemetry-specific coreVersion override would have made the
CLI product version a third, contradictory source of core_version (the
package walk already yields 0.2.0 for source-layout hosts). Instead
_base/version.ts becomes the single owner: a bundler define for hosts that
inline the source (Electron main), the package.json walk moved here from
app/telemetry/coreVersion.ts as the fallback, 'unknown' last resort.
CloudAppender and the MCP client version now read the same function, which
also lifts MCP's client version off the 0.0.0 stub.
